✨ Refactor imports: update service module paths for consistency
