+++ services/catalog/src/simcore_service_catalog/core/background_tasks.py
@@ -23,7 +23,7 @@ from models_library.services_types import ServiceKey, ServiceVersion
 from packaging.version import Version
 from pydantic import ValidationError
 from simcore_service_catalog.api._dependencies.director import get_director_api
-from simcore_service_catalog.services import manifest
+from simcore_service_catalog.service import manifest
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncEngine
 
@@ -31,7 +31,7 @@ from ..db.repositories.groups import GroupsRepository
 from ..db.repositories.projects import ProjectsRepository
 from ..db.repositories.services import ServicesRepository
 from ..models.services_db import ServiceAccessRightsAtDB, ServiceMetaDataDBCreate
-from ..services import access_rights
+from ..service import access_rights
 
 _logger = logging.getLogger(__name__)
