+++ services/catalog/src/simcore_service_catalog/service/access_rights.py
