+++ services/catalog/src/simcore_service_catalog/service/rabbitmq.py
