+++ services/catalog/tests/unit/with_dbs/test_services_access_rights.py
@@ -11,7 +11,7 @@ from models_library.services import ServiceMetaDataPublished, ServiceVersion
 from pydantic import TypeAdapter
 from simcore_service_catalog.db.repositories.services import ServicesRepository
 from simcore_service_catalog.models.services_db import ServiceAccessRightsAtDB
-from simcore_service_catalog.services.access_rights import (
+from simcore_service_catalog.service.access_rights import (
     evaluate_auto_upgrade_policy,
     evaluate_default_policy,
     reduce_access_rights,
