+++ services/catalog/tests/unit/with_dbs/test_services_services_api.py
@@ -19,8 +19,8 @@ from respx.router import MockRouter
 from simcore_service_catalog.api._dependencies.director import get_director_api
 from simcore_service_catalog.db.repositories.groups import GroupsRepository
 from simcore_service_catalog.db.repositories.services import ServicesRepository
-from simcore_service_catalog.services import manifest, services_api
-from simcore_service_catalog.services.director import DirectorApi
+from simcore_service_catalog.service import manifest, services_api
+from simcore_service_catalog.service.director import DirectorApi
 from sqlalchemy.ext.asyncio import AsyncEngine
 
 pytest_simcore_core_services_selection = [
